Rename and split DebugView into sections
