+++ ScoutIP/Components/Debug/DebugView.swift
@@ -1,0 +1,20 @@
+//
+// Copyright 2026 Mikhail Kasianov
+//
+// Use of this source code is governed by an MIT-style
+// license that can be found in the LICENSE file or at
+// https://opensource.org/licenses/MIT.
+
+import SwiftUI
+
+struct DebugView: View {
+  var body: some View {
+    List {
+      LoggingSection()
+      MetricsSection()
+      StressSection()
+      CrashSection()
+    }
+    .navigationTitle("Debug")
+  }
+}
